Refactor Fronts ad position functions
